Replaced unsigned with the correct integral type (which resolved some warnings)

+++ Arrangement_on_surface_2/examples/Arrangement_on_surface_2/dcel_extension.cpp
@@ -8,8 +8,8 @@
 
 enum Color {BLUE, RED, WHITE};
 
-typedef CGAL::Arr_extended_dcel<Traits, Color, bool, unsigned int>    Dcel;
-typedef CGAL::Arrangement_2<Traits, Dcel>                   Ex_arrangement;
+typedef CGAL::Arr_extended_dcel<Traits, Color, bool, size_t> Dcel;
+typedef CGAL::Arrangement_2<Traits, Dcel>                    Ex_arrangement;
 
 int main() {
   // Construct the arrangement containing two intersecting triangles.
